+++ lib/src/comet2/device.dart
@@ -3,11 +3,7 @@ typedef Output = void Function(String);
 
 /// Interface.
 class Device {
-  final Input input = () {
-    return 'hello, world';
-  };
+  Input input = () => 'hello, world';
 
-  final Output output = (final String s) {
-    print(s);
-  };
+  Output output = (final String s) => print(s);
 }