+++ lib/src/comet2/supervisor_call/util.dart
@@ -1,1 +1,1 @@
-const eof = 0xffff;
+const eof = -1;