+++ test/comet2/instruction/add_arithmetic_test.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:tiamat/src/comet2/instruction.dart';
+import 'package:tiamat/src/comet2/instruction/instruction.dart';
 import 'package:tiamat/src/comet2/resource.dart';
 import 'package:test/test.dart';