+++ web/src/app.dart
@@ -8,12 +8,12 @@ import './component/content_box.dart';
 import './component/editor.dart';
 import './component/information.dart';
 
-const version = '0.1.3+nullsafety';
+const version = '0.2.0+nullsafety';
 
 Element app() {
-  final cc = Casl2();
-  final r = Resource();
-  final c = Comet2();
+  final casl2 = Casl2();
+  final comet2 = Comet2();
+  final r = comet2.resource;
 
   final inputValues = <String>[];
 
@@ -21,7 +21,7 @@ Element app() {
   final editor = Editor(asm);
   final output = TextAreaElement()..disabled = true;
   final input = TextAreaElement();
-  final control = ControlPanel(r, c, cc, onPreExecute: () {
+  final control = ControlPanel(r, comet2, casl2, onPreExecute: () {
     inputValues.clear();
     inputValues.addAll((input.value ?? '').split('\n'));
     r.PR = 0;
@@ -46,11 +46,11 @@ Element app() {
     output.value = '';
   });
 
-  c.sv
-    ..write = (s) {
+  comet2.device
+    ..output = (s) {
       output.value = '${output.value ?? ''}$s\n';
     }
-    ..read = () {
+    ..input = () {
       final s = inputValues;
       var x = 0;
       return () => s[x++ % s.length];